+++ .gitignore
@@ -3,6 +3,7 @@ __pycache__/
 *.py[cod]
 *$py.class
 output/*
+*.swp
 
 # C extensions
 *.so
