+++ README.md
@@ -1,13 +1,27 @@
-# analysis
-tools for analyzing web scraper data
+# LDA
+LDA algorithm, converts raw scraper data into term-document matrix
 
-##Installation
+# Installation
+
+## Third-Party
+
+```
+pip install git+https://github.com/BIDS-projects/lda.git
+```
+
+## Development
 
 1. Setup a virtual environment.
 2. `pip install -r requirements.txt`.
 
 ##Test
+
 ```
 py.test tests
 ```
-# gj
+
+# Usage
+
+```
+import bidslda
+```
