+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup, find_packages
+
+
+setup(
+    name='lda',
+    author='Louie Lee, Don Han, Alvin Wan',
+    version='0.0.1',
+    description='BIDS LDA converting raw scraper data into term-document matrix',
+    packages=find_packages()
+)
